Ensure ShapedLikeNDArray is not iterable for scalars.
