+++ CHANGES.rst
@@ -137,7 +137,8 @@ API Changes
   - ``SkyCoord``, ``ICRS``, and other coordinate objects, as well as the
     underlying representations such as ``SphericalRepresentation`` and
     ``CartesianRepresentation`` can now be reshaped using methods named like the
-    numpy ones for ``ndarray`` (``reshape``, ``swapaxes``, etc.) [#4123, #5254]
+    numpy ones for ``ndarray`` (``reshape``, ``swapaxes``, etc.)
+    [#4123, #5254, #5482]
 
   - The ``obsgeoloc`` and ``obsgeovel`` attributes of ``GCRS`` and
     ``PrecessedGeocentric`` frames are now stored and returned as
